bunch of small optimizations
