+++ examples/CMakeLists.txt
@@ -8,10 +8,10 @@ add_example(stereo-scope)
 add_example(ranged-spectrum)
 add_example(polar-spectrum)
 add_example(stereo-polar-spectrum)
-add_example(shake-bass)
+# add_example(shake-bass)
 add_example(old-bass-nation)
 
-if(AUDIOVIZ_USE_IMGUI)
-	add_example(spectrum-imgui)
-	add_example(scope-imgui)
-endif()
+# if(AUDIOVIZ_USE_IMGUI)
+# 	add_example(spectrum-imgui)
+# 	add_example(scope-imgui)
+# endif()
