+++ examples/stereo-polar-spectrum.cpp
@@ -1,6 +1,6 @@
 #include <audioviz/Base.hpp>
-#include <audioviz/SpectrumDrawable_new.hpp>
-#include <audioviz/fft/AudioAnalyzer_new.hpp>
+#include <audioviz/SpectrumDrawable.hpp>
+#include <audioviz/fft/AudioAnalyzer.hpp>
 #include <audioviz/fft/FrequencyAnalyzer.hpp>
 #include <audioviz/fft/Interpolator.hpp>
 #include <audioviz/fx/Polar.hpp>
@@ -36,7 +36,7 @@ struct StereoPolarSpectrum : audioviz::Base
 	audioviz::ColorSettings cs;
 	audioviz::SpectrumDrawable_new spectrum;
 	audioviz::FrequencyAnalyzer fa;
-	audioviz::AudioAnalyzer_new aa{sample_rate_hz, fft_size};
+	audioviz::AudioAnalyzer aa{sample_rate_hz, fft_size};
 	audioviz::Interpolator ip;
 
 	std::span<const float> audio_buffer;
